Jet-tau fake rate: added a matching to generated taus
In the GenParticleCollection I find the final tau (status=2, and incorrectly = 1 -- TOFIX).
The sum of p4 of non-neutrino daughters is taken as "visible tau".
And added to the visible_gen_taus vector<LorentzVector>.
This collection is used for dR matching with jets in record_jets_fakerate function.
If a jet has partonFlavour = 0 and matches to tau (dR<1)
then its' jet_origin := 15 and recorded as such.

	modified:   eventSelection.cc
	modified:   generationTree.cc
	modified:   jetSelectionForTauFakeRate.cc
	modified:   jetSelectionForTauFakeRate.h