Jet-tau fake rate: non-exclusive HLTs (no separate jetmu); up jet origins stack distr
Now HLTs are like QCD/WJets selection: if event passes HLT_jet it is recorded with HLTjet,
if it also passes HLT_Muon -- it is also recorded with HLTmu.

In jobing the stack distrs for jet origins takes --normalize and --verbose options.
It normalizes the overall integral of the MC stack to Data.

	modified:   jetSelectionForTauFakeRate.cc
	modified:   jobing/dtag_xsecs.h
	modified:   jobing/jet_distrs_stacked_per_origin.cc