Jet-tau fake rates: fixing of trigger matching in MC
Now with variable for trigger names "HLT_PFJet40_v" is used as jetHLT name,
and the wild card "*" is added at the trigger bits call.
The matching to trigger objects is done without "*" -- as in triggerMatching.cc script.

The jobs with Summer16 MC which have the pt bump (at around 50GeV?) don't have issues with this
--- the trigger-name "HLT_PFJet40" was hard coded at the time.

Continue investigation the stuff.

The jet_distrs stack by origins is up with SetLogy for x (pT) and z (Radius) projections.
It doesn't work that well actually -- should make it an option.

	modified:   jetSelectionForTauFakeRate.cc
	modified:   jobing/jet_distrs_stacked_per_origin.cc
	modified:   triggerMatching.cc