Shoved the record functions with jetDistr records into a separate -header- file, Added Processing Muons with ID_ISO_Kinematics procedure
The functions are used as a header file now, while I'm figuring out how scram b compiles the project.
(Apparently it makes a shared lib for the whole project? Dosn't compile separate files into a binary statically.)
The ProcessingMuons.cc is also included as header for now.
After other particles processing is done and all the processing functions are in separate files,
I will need to decople them into seprates -source- files for convenient compilation.

For now it's just a clean-up. Which should be followed with an update of all the ID/ISO POG recommendations.
The goal is the ttbar cut-based weightflow for Summer16+ReReco

	new file:   ProcessingMuons.cc
	modified:   demoKinoTTbar.cc
	modified:   easyKinoTTbar.cc
	modified:   eventSelection.cc
	deleted:    eventSelection_newTauMVAbasedBtag.cc
	modified:   generationTree.cc
	modified:   jetSelectionForTauFakeRate.cc
	deleted:    jetSelectionForTauFakeRate.h
	modified:   printoutTriggers.cc
	new file:   recordFuncs.h
	modified:   tauOrigins.cc
	modified:   triggerMatching.cc